(Rigoberto) Hotel schema checked
